change f64 to decimal

+++ host/src/main.rs
@@ -21,9 +21,9 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     // input : [a0, a1, ..., an] for a0 + a1 * x + ... + an * x^n
     let n: usize = 2;
-    let ax: Vec<i64> = vec![-1, 1]; // -1 + x
+    let ax: Vec<f64> = vec![-1., 1.]; // -1 + x
     let m: usize = 5;
-    let bx: Vec<i64> = vec![1, 1, 1, 1, 1]; // 1 + x + x^2 + x^3 + x^4
+    let bx: Vec<f64> = vec![1., 1., 1., 1., 1.]; // 1 + x + x^2 + x^3 + x^4
 
     let env = ExecutorEnv::builder()
         .write(&n)?
